requirements.txt
# Core app deps
fastapi==0.128.4
uvicorn[standard]==0.40.0
python-dotenv==1.2.1
watchfiles==1.1.1

# RAG stack
chromadb==1.4.1
sentence-transformers==5.2.2
pypdf==6.7.0

# LLM / orchestration
google-adk==1.24.1
openai==2.17.0
litellm>=1.81.9

# Optional: for CLI niceties
typer==0.12.5
